# Conflicts:
#	Explorer/Assets/DCL/Multiplayer/Profiles/Announcements/RemoteAnnouncement.cs
#	Explorer/Assets/DCL/Multiplayer/Profiles/RemoteProfiles/RemoteProfiles.cs

+++ Explorer/Assets/DCL/Landscape/Utils/TerrainGeneratorLocalCache.cs
@@ -80,18 +80,8 @@ namespace DCL.Landscape.Utils
         {
             var path = GetFilePath(seed, chunkSize, version);
             checksum = parcelChecksum;
-
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-                var dictionaryPath = GetDictionaryDirectory();
-                if (Directory.Exists(dictionaryPath))
-                    Directory.Delete(GetDictionaryDirectory(), true);
-            }
-            
             using FileStream fileStream = File.Create(path);
             FORMATTER.Serialize(fileStream, this);
-
         }
 
         public void SaveArrayToFile<T>(string name, string offsetX, string offsetZ, T[] arrayToSave) where T : struct
@@ -151,36 +141,48 @@ namespace DCL.Landscape.Utils
                 checksum = parcelChecksum,
             };
 
-            var path = GetFilePath(seed, chunkSize, version);
+            string? filePath = GetFilePath(seed, chunkSize, version);
             var dictionaryPath = GetDictionaryDirectory();
 
-            if (force && File.Exists(path))
-            {
-                File.Delete(path);
-                if (Directory.Exists(dictionaryPath))
-                    Directory.Delete(GetDictionaryDirectory(), true);
-            }
+            CheckCorruptStates();
 
-            if (!File.Exists(path))
+            if (force && File.Exists(filePath))
+                ClearCache();
+
+            if (!File.Exists(filePath))
             {
-                //If the file does not exist, but the dictionary path does, theres a possible corrupt state. 
-                //We gotta clear it
-                if (Directory.Exists(dictionaryPath))
-                    Directory.Delete(dictionaryPath, true);
                 Directory.CreateDirectory(dictionaryPath);
-                
                 return emptyCache;
             }
 
-            await using var fileStream = new FileStream(path, FileMode.Open);
+            await using var fileStream = new FileStream(filePath, FileMode.Open);
 
             TerrainLocalCache? localCache = await UniTask.RunOnThreadPool(() => (TerrainLocalCache)FORMATTER.Deserialize(fileStream));
 
             if (localCache.checksum != parcelChecksum)
+            {
+                ClearCache();
+                Directory.CreateDirectory(dictionaryPath);
                 return emptyCache;
+            }
 
             localCache.isValid = true;
             return localCache;
+
+            void CheckCorruptStates()
+            {
+                if (File.Exists(filePath) && !Directory.Exists(dictionaryPath))
+                    File.Delete(filePath);
+
+                if (!File.Exists(filePath) && Directory.Exists(dictionaryPath))
+                    Directory.Delete(dictionaryPath, true);
+            }
+
+            void ClearCache()
+            {
+                File.Delete(filePath);
+                Directory.Delete(dictionaryPath, true);
+            }
         }
 
         public bool IsValid() =>
